refactor(pages): Move victim memory page to its module #9
