+++ src/memories/pages/index.js
@@ -1,0 +1,1 @@
+export { default as MemoryVictim } from './memory-victim'
