+++ src/memories/pages/memory-victim/index.js
@@ -1,0 +1,1 @@
+export { default } from './page.connected'
