+++ src/memories/pages/memory-victim/page.connected.js
@@ -1,7 +1,7 @@
 import { connect } from 'preact-redux'
 import { graphql } from 'react-apollo'
-import * as queries from '../../graphql/queries'
-import MemoryVictim from './memory-victim'
+import * as queries from '../../../graphql/queries'
+import MemoryVictim from './page'
 
 const mapStateToProps = (state, props) => ({
   memory: props.data.loading ? {} : props.data.memoryById,
